initial draft of the resolver and big refactor of many parts - Add default imports in parse - Refactor some parts of the scope - Scope lookups are kind-based now - Improved test cases for full parse with resolution - Refactor of some parts of the AST
The resolver right now is a mess, and requires a little bit of love.
For the moment, it's working and it's good enough. Whenever there's time
it should be top priority to split some of its functions into smaller ones
and improve tests cases. Right now, there are more pressing needs.